Upgraded to gradle-plugin-cq-bundle v2.1.0
Brought over a few functions that used to be in -bundle v.2.0 and put them in CqPackageHelper such as
  * doAcrossServers
  * startInactiveBundles
  * validateBundles
  * validateRemoteBundles
  * validateAllBundles
  * hasAnInactiveBundle
  * uninstallBundles
  * uninstallAllBundles

Removed the "checkBundleStatus" since it's redundant with "validateBundles"

Changed a number of methods in CqPackageHelper to return their status instead of 'void'. Still need to change CqPackagePlugin and the like to check for the status.