A lot of cleanup to the task definitions
Renamed the tasks to be less generic (e.g., "upload" -> "uploadPackage")

Removed tasks that are redundant or are of marginal utility.

Added stronger typing to task dependencies in plugin declaration, as well as more javadocs

Changed the actual task declarations to be consistent in the form of declaration.

Fixes GH-13